Trying to add individual event page and bg color
Code to go to individual event page on click and also change background
color of app to white. I have yet to test this code because the server
with the timeline info randomly just stopped letting me connect. I also
added an error message to show in the place of the listview if the data
fails to load